refactor(hall-of-fame): 컴포넌트 분리 및 코드 정리 (#28)
- VoteButton tv() variant 적용
- VoteCard '자세히 보기' onClick 연결
- RANKING_PERIOD_CONFIG로 삼항연산자 통합
- EntryDetailModal 내부 컴포넌트 파일 분리
