+++ src/app/(main)/hall-of-fame/_ui/EntryDescription.tsx
@@ -1,0 +1,12 @@
+import type { ContestEntry } from '@/shared/types'
+
+/** 하단 설명 */
+const EntryDescription = ({ entry }: { entry: ContestEntry }) => (
+  <div className="shrink-0 bg-[#2f2f2f] px-5 py-3 tab:px-16 tab:py-8">
+    <p className="text-sm font-medium leading-relaxed text-white">
+      {entry.description}
+    </p>
+  </div>
+)
+
+export { EntryDescription }
