+++ src/app/(main)/hall-of-fame/_ui/EntryHeader.tsx
@@ -1,0 +1,32 @@
+import { VoteIcon } from '@/shared/assets/icons'
+import type { ContestEntry } from '@/shared/types'
+
+interface EntryHeaderProps {
+  entry: ContestEntry
+  onClose: () => void
+  className?: string
+}
+
+/** 상단 헤더 (닫기 + 투표 상태) */
+const EntryHeader = ({ entry, onClose, className }: EntryHeaderProps) => (
+  <div className={`flex shrink-0 items-center ${className ?? ''}`}>
+    <button
+      type="button"
+      className="flex w-[3.625rem] items-center justify-center rounded-full border border-[#d4d4d4] p-[0.625rem] tab:w-[5.75rem]"
+      onClick={onClose}
+    >
+      <span className="text-sm font-medium text-white">닫기</span>
+    </button>
+    {entry.isVoted && (
+      <div className="ml-auto flex items-center gap-1.5 tab:ml-0 tab:gap-2">
+        <span className="text-sm font-medium text-white tab:text-base">투표했습니다</span>
+        <VoteIcon className="size-6 text-white" />
+        <span className="text-sm font-semibold leading-[1.375rem] text-white">
+          {entry.voteCount}
+        </span>
+      </div>
+    )}
+  </div>
+)
+
+export { EntryHeader }
