+++ src/app/(main)/hall-of-fame/_ui/HallOfFameContent.tsx
@@ -17,13 +17,18 @@ import { VoteCard } from './VoteCard'
 
 type RankingPeriod = 'current' | 'previous'
 
+const RANKING_PERIOD_CONFIG: Record<RankingPeriod, { title: string; buttonLabel: string; next: RankingPeriod }> = {
+  current: { title: '실시간 랭킹 1위~3위', buttonLabel: '< 저번주 랭킹', next: 'previous' },
+  previous: { title: '저번주 랭킹 1위~3위', buttonLabel: '이번주 랭킹>', next: 'current' },
+}
+
 const HallOfFameContent = () => {
   const [isRankingOpen, setIsRankingOpen] = useState(true)
   const [isVoteOpen, setIsVoteOpen] = useState(true)
   const [rankingPeriod, setRankingPeriod] = useState<RankingPeriod>('current')
   const [selectedEntry, setSelectedEntry] = useState<ContestEntry | null>(null)
 
-  const isCurrent = rankingPeriod === 'current'
+  const periodConfig = RANKING_PERIOD_CONFIG[rankingPeriod]
 
   return (
     <div className="flex w-full flex-col pb-12">
@@ -49,7 +54,7 @@ const HallOfFameContent = () => {
       <Container className="mt-6 tab:mt-16">
         <div className="flex flex-col gap-3 tab:gap-5">
           <SectionHeader
-            title={isCurrent ? '실시간 랭킹 1위~3위' : '저번주 랭킹 1위~3위'}
+            title={periodConfig.title}
             subtitle={MOCK_CONTEST_INFO.dateRange}
             collapsible
             collapsed={!isRankingOpen}
@@ -58,10 +63,10 @@ const HallOfFameContent = () => {
               <button
                 type="button"
                 className="flex h-9 items-center justify-center rounded-full bg-[#a4a4a4] px-5"
-                onClick={() => setRankingPeriod(isCurrent ? 'previous' : 'current')}
+                onClick={() => setRankingPeriod(periodConfig.next)}
               >
                 <span className="whitespace-nowrap text-sm font-semibold text-white tab:text-base">
-                  {isCurrent ? '< 저번주 랭킹' : '이번주 랭킹>'}
+                  {periodConfig.buttonLabel}
                 </span>
               </button>
             }
