+++ src/app/(main)/hall-of-fame/_ui/VoteButton.tsx
@@ -1,6 +1,18 @@
 'use client'
 
+import { tv } from 'tailwind-variants'
 import { VoteIcon } from '@/shared/assets/icons'
+
+const voteButtonVariants = tv({
+  base: 'flex items-center justify-center gap-[0.625rem] rounded-full px-[0.625rem] py-1',
+  variants: {
+    voted: {
+      true: 'bg-[#a4a4a4]',
+      false: 'bg-[#5d5d5d]',
+    },
+  },
+  defaultVariants: { voted: false },
+})
 
 interface VoteButtonProps {
   isVoted: boolean
@@ -13,9 +25,7 @@ const VoteButton = ({ isVoted, voteCount, label, className }: VoteButtonProps) =
   return (
     <button
       type="button"
-      className={`flex items-center justify-center gap-[0.625rem] rounded-full px-[0.625rem] py-1 ${
-        isVoted ? 'bg-[#a4a4a4]' : 'bg-[#5d5d5d]'
-      } ${className ?? ''}`}
+      className={voteButtonVariants({ voted: isVoted, className })}
     >
       <VoteIcon className="size-6 text-white" />
       <span className="text-sm font-semibold leading-[1.375rem] text-white">
